+++ client/apps/eternum-mobile/src/app/config/config.ts
