+++ client/apps/eternum-mobile/src/features/resource-balances/model/use-resource-balances.ts
@@ -1,4 +1,4 @@
-import { useBlockTimestamp } from "@/shared/lib/hooks/use-block-timestamp";
+import { useBlockTimestamp } from "@/shared/hooks/use-block-timestamp";
 import { ID, resources } from "@bibliothecadao/eternum";
 import { useResourceManager } from "@bibliothecadao/react";
 import { useCallback, useEffect, useState } from "react";
