+++ client/apps/eternum-mobile/src/features/resources-production/model/use-production.ts
@@ -1,4 +1,4 @@
-import { getBlockTimestamp } from "@/shared/lib/hooks/use-block-timestamp";
+import { getBlockTimestamp } from "@/shared/hooks/use-block-timestamp";
 import useStore from "@/shared/store";
 import { getProducedResource, ResourcesIds } from "@bibliothecadao/eternum";
 import {
