+++ client/apps/eternum-mobile/src/features/upgrade-structure/model/use-structure-upgrade.ts
@@ -1,4 +1,4 @@
-import { getBlockTimestamp } from "@/shared/lib/hooks/use-block-timestamp";
+import { getBlockTimestamp } from "@/shared/hooks/use-block-timestamp";
 import {
   configManager,
   divideByPrecision,
