+++ client/apps/eternum-mobile/src/pages/realm/ui/components/production-widgets-section.tsx
@@ -1,4 +1,4 @@
-import { getBlockTimestamp } from "@/shared/lib/hooks/use-block-timestamp";
+import { getBlockTimestamp } from "@/shared/hooks/use-block-timestamp";
 import { ProductionWidget } from "@/widgets/production-widget";
 import { LaborBuilding } from "@/widgets/production-widget/model/types";
 import { Building, getProducedResource, RealmInfo } from "@bibliothecadao/eternum";
