+++ client/apps/eternum-mobile/src/shared/hooks/use-block-timestamp.ts
