+++ client/apps/eternum-mobile/src/widgets/production-widget/ui/resources-production-drawer.tsx
@@ -1,5 +1,5 @@
 import { getLaborConfig } from "@/features/resources-production/lib/labor";
-import { getBlockTimestamp } from "@/shared/lib/hooks/use-block-timestamp";
+import { getBlockTimestamp } from "@/shared/hooks/use-block-timestamp";
 import { Button } from "@/shared/ui/button";
 import { Card, CardContent } from "@/shared/ui/card";
 import { Drawer, DrawerContent, DrawerHeader, DrawerTitle } from "@/shared/ui/drawer";
