+++ client/apps/eternum-mobile/src/widgets/upgrade-castle/ui/upgrade-drawer.tsx
@@ -1,5 +1,5 @@
 import { useStructureUpgrade } from "@/features/upgrade-structure";
-import { getBlockTimestamp } from "@/shared/lib/hooks/use-block-timestamp";
+import { getBlockTimestamp } from "@/shared/hooks/use-block-timestamp";
 import { Button } from "@/shared/ui/button";
 import { Drawer, DrawerContent, DrawerDescription, DrawerHeader, DrawerTitle } from "@/shared/ui/drawer";
 import { ResourceIcon } from "@/shared/ui/resource-icon";
